refactor

+++ zfs/manifests/zfs_create.pp
@@ -1,26 +1,29 @@
 define zfs::zfs_create (
   $ensure = present,
   $filesystem = '',
+  $noop = false,
 ){
   Exec {
     path => '/sbin:/usr/sbin:/usr/local/sbin:/bin:/usr/bin:/usr/local/bin'
   }
 
-  case $ensure {
-    present: {
-      exec { "${name}_zfs_create":
-        command => "zfs create ${name}",
-        unless  => "zfs list ${name}"
-      }
-      exec { "zfs set mountpoint=${filesystem} ${name}":
-        unless => "zfs get mountpoint ${name} | grep -q ' ${filesystem} '",
+  if !$noop {
+    case $ensure {
+      present: {
+        exec { "${name}_zfs_create":
+          command => "zfs create ${name}",
+          unless  => "zfs list ${name}"
+        }
+        exec { "zfs set mountpoint=${filesystem} ${name}":
+          unless => "zfs get mountpoint ${name} | grep -q ' ${filesystem} '",
 
-        require => Exec["${name}_zfs_create"],
+          require => Exec["${name}_zfs_create"],
+        }
       }
-    }
-    absent: {
-      exec { "zfs destroy -r ${name}":
-        onlyif => "zfs list ${name}"
+      absent: {
+        exec { "zfs destroy -r ${name}":
+          onlyif => "zfs list ${name}"
+        }
       }
     }
   }
